rearrange how/where we spawn API endpoints
